[FEAT] Split updateSkills into its own file with shared cost helper
Extracts updateSkills and UpdateSkillsAction out of choseSkill.ts into
their own file with matching tests, and introduces a shared
chargeSkillEdit helper in resetSkills.ts that drives the free/gems/
ticket cost for both resetSkills and updateSkills so the two paths
cannot drift in cost.

Also rejects no-op skill updates where the sanitised desired skill set
equals the current one, so a client resending the existing build no
longer burns the free cooldown / gems / ticket.
