+++ src/features/game/events/landExpansion/resetSkills.ts
@@ -1,6 +1,6 @@
 import { getKeys } from "lib/object";
 import type { GameState } from "features/game/types/game";
-import { produce } from "immer";
+import { produce, Draft } from "immer";
 import Decimal from "decimal.js-light";
 import { populateSaltFarm } from "features/game/types/salt";
 
@@ -46,6 +46,74 @@ export function canResetForFree(
   return timeUntilNextReset <= 0;
 }
 
+// Shared by resetSkills and updateSkills so the two paths cannot drift in cost.
+export function chargeSkillEdit({
+  game,
+  paymentType,
+  createdAt,
+}: {
+  game: Draft<GameState>;
+  paymentType: PaymentType;
+  createdAt: number;
+}) {
+  const { bumpkin } = game;
+
+  if (bumpkin == undefined) {
+    throw new Error("You do not have a Bumpkin!");
+  }
+
+  const { paidSkillResets = 0, previousFreeSkillResetAt = 0 } = bumpkin;
+
+  switch (paymentType) {
+    case "free": {
+      if (!canResetForFree(previousFreeSkillResetAt, createdAt)) {
+        const timeToNextFreeResetInMilliseconds = getTimeUntilNextFreeReset(
+          previousFreeSkillResetAt,
+          createdAt,
+        );
+        const daysRemaining = Math.ceil(
+          timeToNextFreeResetInMilliseconds / (24 * 60 * 60 * 1000),
+        );
+        throw new Error(
+          `Wait ${daysRemaining} more days for free reset or use gems`,
+        );
+      }
+
+      bumpkin.paidSkillResets = 0;
+      bumpkin.previousFreeSkillResetAt = createdAt;
+      break;
+    }
+    case "gems": {
+      const gemCost = getGemCost(paidSkillResets);
+      const gemBalance = game.inventory.Gem ?? new Decimal(0);
+
+      if (gemBalance.lt(gemCost)) {
+        throw new Error(`Not enough gems. Cost: ${gemCost} gems`);
+      }
+
+      game.inventory.Gem = gemBalance.minus(gemCost);
+      bumpkin.paidSkillResets = paidSkillResets + 1;
+      break;
+    }
+    case "ticket": {
+      const ticketBalance =
+        game.inventory["Skill Reset Ticket"] ?? new Decimal(0);
+
+      if (ticketBalance.lt(1)) {
+        throw new Error("You do not have a Skill Reset Ticket");
+      }
+
+      game.inventory["Skill Reset Ticket"] = ticketBalance.minus(1);
+      bumpkin.paidSkillResets = paidSkillResets + 1;
+      break;
+    }
+    default: {
+      const _exhaustive: never = paymentType;
+      throw new Error(`Unknown payment type: ${_exhaustive}`);
+    }
+  }
+}
+
 export function resetSkills({
   state,
   action,
@@ -53,76 +121,16 @@ export function resetSkills({
 }: Options) {
   return produce(state, (game) => {
     const { bumpkin } = game;
-    const {
-      paidSkillResets = 0,
-      previousFreeSkillResetAt = 0,
-      skills,
-    } = bumpkin;
 
-    // Check if bumpkin exists
     if (bumpkin == undefined) {
       throw new Error("You do not have a Bumpkin!");
     }
 
-    // Check if bumpkin has any skills
-    if (getKeys(skills).length === 0) {
+    if (getKeys(bumpkin.skills).length === 0) {
       throw new Error("You do not have any skills to reset");
     }
 
-    const { paymentType } = action;
-    let didPaidReset = false;
-
-    switch (paymentType) {
-      case "free": {
-        if (!canResetForFree(previousFreeSkillResetAt, createdAt)) {
-          const timeToNextFreeResetInMilliseconds = getTimeUntilNextFreeReset(
-            previousFreeSkillResetAt,
-            createdAt,
-          );
-          const daysRemaining = Math.ceil(
-            timeToNextFreeResetInMilliseconds / (24 * 60 * 60 * 1000),
-          );
-          throw new Error(
-            `Wait ${daysRemaining} more days for free reset or use gems`,
-          );
-        }
-
-        bumpkin.paidSkillResets = 0;
-        bumpkin.previousFreeSkillResetAt = createdAt;
-        break;
-      }
-      case "gems": {
-        const gemCost = getGemCost(paidSkillResets);
-
-        if (game.inventory.Gem?.lt(gemCost)) {
-          throw new Error(`Not enough gems. Cost: ${gemCost} gems`);
-        }
-
-        game.inventory.Gem = game.inventory.Gem?.minus(gemCost);
-        didPaidReset = true;
-        break;
-      }
-      case "ticket": {
-        const ticketBalance =
-          game.inventory["Skill Reset Ticket"] ?? new Decimal(0);
-
-        if (ticketBalance.lt(1)) {
-          throw new Error("You do not have a Skill Reset Ticket");
-        }
-
-        game.inventory["Skill Reset Ticket"] = ticketBalance.minus(1);
-        didPaidReset = true;
-        break;
-      }
-      default: {
-        const _exhaustive: never = paymentType;
-        throw new Error(`Unknown payment type: ${_exhaustive}`);
-      }
-    }
-
-    if (didPaidReset) {
-      bumpkin.paidSkillResets = paidSkillResets + 1;
-    }
+    chargeSkillEdit({ game, paymentType: action.paymentType, createdAt });
 
     bumpkin.skills = {};
